chore(sign): move old files
